langchain/harmful-advice/requirements.txt
langchain==0.0.352
langchain-openai
chromadb==0.4.7
python-dotenv==1.0.0
tiktoken
requests==2.31.0
Markdown==3.5
unstructured[all-docs]
